+++ apps/api/src/modules/recipes/recipes.module.ts
@@ -1,7 +1,9 @@
 import { Module } from '@nestjs/common';
 import { RecipeService } from './recipe.service';
+import { RecipesController } from './recipes.controller';
 
 @Module({
+  controllers: [RecipesController],
   providers: [RecipeService],
   exports: [RecipeService],
 })